# snow-x-ui/src/lib/api/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "snow-x-backend",
		"description": "A soft and simple blogging platform for everyone.",
		"contact": { "name": "Levi Lim", "email": "shiueo.csh@gmail.com" },
		"license": { "name": "" },
		"version": "0.8.4"
	},
	"paths": {
...
		"/v0/post/edit": {
			"post": {
				"tags": ["Post"],
				"operationId": "get_post_for_edit",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/GetPostForEditRequest" } } },
					"required": true
				},
				"responses": {
					"200": {
						"description": "Post retrieved for editing successfully",
						"content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostEditInfoResponse" } } }
					},
					"400": { "description": "Invalid input" },
					"404": { "description": "Post not found" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"GetPostForEditRequest": {
				"type": "object",
				"required": ["slug"],
				"properties": { "slug": { "type": "string" } }
			},
...
			"PostEditInfoResponse": {
				"type": "object",
				"required": ["id", "title", "content", "created_at", "slug", "tags"],
				"properties": {
					"content": { "type": "string" },
					"created_at": { "type": "string", "format": "date-time" },
					"id": { "type": "string", "format": "uuid" },
					"slug": { "type": "string" },
					"summary": { "type": ["string", "null"] },
					"tags": { "type": "array", "items": { "type": "string" } },
					"title": { "type": "string" },
					"updated_at": { "type": ["string", "null"], "format": "date-time" }
				}
			},
...
		},
		"securitySchemes": {
			"anonymous_id_cookie": { "type": "apiKey", "in": "cookie", "name": "anonymous_user_id" },
			"bearer_auth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
			"refresh_token_cookie": { "type": "apiKey", "in": "cookie", "name": "refresh_token" }
		}
	},
...
}